+++ resources/views/fields/partials/media-js.blade.php
@@ -24,6 +24,7 @@ var ACCESS_PRIVATE = {{ $access_private }};
 			'is_private[]': ( settings.mediaAccess === ACCESS_PRIVATE ? 1 : 0 ),
 			'_token': $('input[name=_token]').val()
 		},
+		dropZone: '.js-file-upload-{{ $id }}',
 		progressall: function (e, data) {
 			var progress = parseInt(data.loaded / data.total * 100, 10);
 			$(this).parent().nextAll('.js-progress').find('.progress-bar').css(