+++ src/Models/Base.php
@@ -1,5 +1,6 @@
 <?php namespace Bozboz\Admin\Models;
 
+use Bozboz\Admin\Services\Validators\EmptyValidator;
 use Bozboz\Admin\Traits\SanitisesInputTrait;
 use Illuminate\Database\Eloquent\Model;
 
@@ -11,4 +12,14 @@ abstract class Base extends Model implements BaseInterface
 	 * @var array
 	 */
 	protected $nullable = [];
+
+	/**
+	 * Return an empty validator as a default implementation of method
+	 *
+	 * @return Bozboz\Admin\Services\Validators\EmptyValidator
+	 */
+	public function getValidator()
+	{
+		return new EmptyValidator;
+	}
 }